+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/pages/restore.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { hot } from 'react-hot-loader'
 import { useHistory } from 'react-router-dom'
 import { AsyncTasks } from '../../js/model/AsyncTask/async-task'
-import { OpenSearch } from './home'
+import { OpenSearch } from './search'
 
 const Open = () => {
   const history = useHistory()
