+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/pages/search.tsx
